backend/requirements.txt
Flask>=2.3.0
Flask-RESTful>=0.3.9
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
werkzeug>=2.3.0
# For Docker SQL Server connection
pymssql>=2.2.0